bring tests into specs2 2.3, scalacheck 1.11 patterns
`haveTheSameElementsAs` seems to have been dropped; instead I’m using `containsTheSameElementsAs`, which is part of `TraversableMatchers` in specs

There also seems to be a change in how specs should be written, in terms of actually writing out the `def is` string. A specs2-specific interpolator seems to exist, `s2`, that takes in-line expectations. I may go through and update everything, once I figure more out about it. The main effect is that the console output is no longer very attractive.