fix DummyResourceSpecs
The fixes are dubious and can be seen in the changes to `DummyResource`. I’d like someone else to take a look at them soon. Having to append the path in the location of a post-as-create result seems flatly wrong to me. The parsing for a `NoBody` in the `parsePutBody` pathway is a hack, but may be passable for this usage.