consolidate to one HttpResourceActor instance
this combines AbstractResourceActor into HttpResourceActor by directly
extending HttpResourceActor. The methods in AbstractResourceActor are
the same, but we lose the typing of ParsedRequest on HttpResourceActor.
I don’t think that’s a big deal, as only auto-ten code was really using
it anyway.

One spec is still failing. It seems the processReceiveTimeout isn’t
triggering properly. Need to debug that next week.